pyproject.toml
[tool.poetry]
name = "langch-gemini"
version = "0.1.0"
description = ""
authors = ["Shuichi Kataoka <shu01k9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain-google-genai = "^1.0.10"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
